APIs/requirements.txt
azure-functions==1.21.3
pyodbc
pyjwt
